Remove all warnings from pips 64 bits main issues where related to usage of int where intptr_t should be used also note the existence of the c99 standard macro PRIdPTR that expands to ld or d for printf output of intptr_t
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@17302 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/alias-classes/conflicts.c
@@ -138,8 +138,8 @@ static bool old_effects_conflict_p( effect eff1, effect eff2 ) {
               if ( unbounded_expression_p( e1 ) || unbounded_expression_p( e2 ) )
                 conflict_p = TRUE;
               else {
-                int i1 = -1;
-                int i2 = -1;
+                intptr_t i1 = -1;
+                intptr_t i2 = -1;
                 bool i1_p = FALSE;
                 bool i2_p = FALSE;
 
@@ -209,8 +209,8 @@ bool array_references_may_conflict_p( list sl1, list sl2 ) {
     if ( unbounded_expression_p( e1 ) || unbounded_expression_p( e2 ) )
       conflict_p = TRUE;
     else {
-      int i1 = -1;
-      int i2 = -1;
+      intptr_t i1 = -1;
+      intptr_t i2 = -1;
       bool i1_p = FALSE;
       bool i2_p = FALSE;
 
@@ -279,8 +279,8 @@ bool variable_references_may_conflict_p( entity v, list sl1, list sl2 ) {
             conflict_p = FALSE;
           }
         } else {
-          int i1 = -1;
-          int i2 = -1;
+          intptr_t i1 = -1;
+          intptr_t i2 = -1;
           bool i1_p = FALSE;
           bool i2_p = FALSE;
