+++ src/Libs/hpfc/hpfc-util.c
@@ -617,7 +617,7 @@ int HpfcExpressionToInt(e)
 expression e;
 {
     normalized n = expression_normalized(e);
-    int val = 0;
+    intptr_t val = 0;
 
     ifdebug(8) print_expression(e);
