+++ src/Libs/phrase/loop_spaghettify.c
@@ -237,7 +237,7 @@ static control make_condition_from_loop (loop the_loop,
 {
   statement condition_statement;
   test condition_test;
-  int increment_value;
+  intptr_t increment_value;
   string intrinsic_name;
   expression test_condition;
