+++ src/Libs/prettyprint/xml_prettyprinter.c
@@ -278,8 +278,8 @@ static string xml_dim_string(list ldim, string name)
 	expression elow = dimension_lower(dim);
 	expression eup = dimension_upper(dim);
 	
-	int low;
-	int up;
+	intptr_t low;
+	intptr_t up;
 	nbdim++;
 	if (expression_integer_value(elow, &low)){
 	  if(nbdim != 1)
@@ -1206,8 +1206,8 @@ for(ld = ldecl; !ENDP(ld); ld = CDR(ld)){
 	
 	 for (dim = variable_dimensions(type_variable(entity_type(var))); !ENDP(dim); dim = CDR(dim)) {
 
-	  int low;
-	  int  up;
+	  intptr_t low;
+	  intptr_t  up;
 	  expression elow = dimension_lower(DIMENSION(CAR(dim)));
 	  expression eup = dimension_upper(DIMENSION(CAR(dim)));
 	  if (expression_integer_value(elow, &low) && expression_integer_value(eup, &up)){
@@ -2571,7 +2571,7 @@ boolean  eval_linear_expression(expression exp, Psysteme ps, int *val)
 
 static void xml_Bounds(expression elow, expression eup,Psysteme prec, string_buffer sb_result)
 {
-  int low,up;
+  intptr_t low,up;
   int valr =0;
   /* Print XML Array LOWER BOUND */
   string_buffer_append_word("LowerBound",sb_result);
@@ -2603,7 +2603,7 @@ static void xml_Bounds(expression elow, expression eup,Psysteme prec, string_buf
 static void xml_Bounds_and_Stride(expression elow, expression eup, expression stride,
 				  Psysteme prec, string_buffer sb_result)
 {
-  int inc;
+  intptr_t inc;
   int  valr =0;
   xml_Bounds(elow, eup,prec,sb_result);
   string_buffer_append_word("Stride",sb_result);
@@ -3198,7 +3198,7 @@ static void  xml_Arguments(statement s, entity function, Pvecteur loop_indices, 
   reference ActualRef=reference_undefined;
   syntax sr;
   effect ef = effect_undefined; 
-  int iexp,ith=0;
+  intptr_t iexp,ith=0;
   int rw_ef=0;
   string aan ="";
   int valr  ;
